+++ backend/src/routes/v0/llm/__init__.py
@@ -30,14 +30,6 @@ router = APIRouter()
                 "application/json": {
                     "example": Answer.model_json_schema()['examples']['new_thread']
                 },
-                # "text/event-stream": {
-                #     "description": "Server-sent events stream",
-                #     "schema": {
-                #         "type": "string",
-                #         "format": "binary",
-                #         "example": 'data: {"event": "ai_chunk", "content": [{"text": "Hello", "type": "text", "index": 0}]}\n\n'
-                #     }
-                # }
             }
         }
     }
@@ -79,7 +71,7 @@ async def chat_completion(
 @router.post(
     "/threads",
     name="Create New Thread",
-    tags=["Thread"],
+    tags=[TAG],
     responses={
         status.HTTP_200_OK: {
             "description": "Latest message from new thread.",
@@ -108,7 +100,8 @@ async def new_thread(
     
     try:
         controller = AgentController(db=db, user_id=user.id if user else None)
-        return await controller.anew_thread(request=request, new_thread=body)
+        output_type = request.headers.get("accept", "application/json")
+        return await controller.query_thread(output_type=output_type, thread=body)
     except httpx.HTTPStatusError as e:
         logger.error(f"Error creating new thread: {str(e)}")
         raise HTTPException(status_code=e.response.status_code , detail=str(e))
@@ -122,7 +115,7 @@ async def new_thread(
 @router.post(
     "/threads/{thread_id}", 
     name="Query Existing Thread",
-    tags=["Thread"],
+    tags=[TAG],
     responses={
         status.HTTP_200_OK: {
             "description": "Latest message from existing thread.",
@@ -151,7 +144,8 @@ async def existing_thread(
 ):
     try:
         controller = AgentController(db=db, user_id=user.id if user else None)
-        return await controller.aexisting_thread(request=request, thread_id=thread_id, existing_thread=body)
+        output_type = request.headers.get("accept", "application/json")
+        return await controller.query_thread(output_type=output_type, thread=body, thread_id=thread_id)
     except httpx.HTTPStatusError as e:
         logger.error(f"Error creating new thread: {str(e)}")
         raise HTTPException(status_code=e.response.status_code , detail=str(e))