+++ frontend/src/hooks/useAgentHook.ts
@@ -75,10 +75,10 @@ export default function useAgentHook() {
 			const response = await createSetting({
 				name: agentDetails.name.replace(/\s+/g, '_').toLowerCase() + ":" + new Date().toISOString().slice(0, 16).replace('T', '_'),
 				value: {
-				system: payload.system,
-				model: payload.model,
-				tools: payload.tools,
-				mcp: payload.mcp
+					system: payload.system,
+					model: payload.model,
+					tools: payload.tools,
+					mcp: payload.mcp
 				}
 			});
 			const settingId = response.data.setting.id;