+++ frontend/src/pages/Home.tsx
@@ -42,7 +42,7 @@ export default function Home() {
                     alt="Logo" 
                     className="w-32 h-32 mx-auto rounded-full" 
                 />
-                <h1 className="text-4xl font-bold mt-2">Ensō Cloud</h1>
+                <h1 className="text-4xl font-bold mt-2">Orchestra</h1>
                 <p className="text-lg mb-2">Powered by <a href="https://github.com/enso-labs/mcp-sse" target="_blank" rel="noopener noreferrer" className="text-primary hover:underline">MCP</a> & <a href="https://github.com/enso-labs/a2a-langgraph" target="_blank" rel="noopener noreferrer" className="text-primary hover:underline">A2A</a></p>
                 <div className="flex flex-row gap-2 mb-2">
                     <a href="https://discord.com/invite/QRfjg4YNzU"><img src="https://img.shields.io/badge/Join-Discord-purple" /></a>